requirements.txt
google-genai
python-dotenv
pillow
opencv-python
numpy
